Changing Procfile parser to work on Windows
