+++ procfile.go
@@ -1,13 +1,11 @@
 package main
 
 import (
-	"bufio"
-	"fmt"
-	_ "github.com/ddollar/forego/Godeps/_workspace/src/github.com/kr/pretty"
-	"io"
 	"math"
 	"os"
 	"regexp"
+
+	_ "github.com/ddollar/forego/Godeps/_workspace/src/github.com/kr/pretty"
 )
 
 var procfileEntryRegexp = regexp.MustCompile("^([A-Za-z0-9_]+):\\s*(.+)$")
@@ -54,19 +52,4 @@ func (pf *Procfile) LongestProcessName(concurrency map[string]int) (longest int)
 		}
 	}
 	return
-}
-
-func parseProcfile(r io.Reader) (*Procfile, error) {
-	pf := new(Procfile)
-	scanner := bufio.NewScanner(r)
-	for scanner.Scan() {
-		parts := procfileEntryRegexp.FindStringSubmatch(scanner.Text())
-		if len(parts) > 0 {
-			pf.Entries = append(pf.Entries, ProcfileEntry{parts[1], parts[2]})
-		}
-	}
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf("Reading Procfile: %s", err)
-	}
-	return pf, nil
 }
